requirements.txt
Django>=5.0
Pillow
qrcode
reportlab
django-crispy-forms
crispy-bootstrap5
asgiref==3.9.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
crispy-bootstrap5==2025.6
distlib==0.3.9
Django==5.2.4
django-crispy-forms==2.5
django-unfold==0.72.0
filelock==3.18.0
idna==3.11
pillow==11.3.0
platformdirs==4.3.8
qrcode==8.2
reportlab==4.4.9
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.3
virtualenv==20.31.2
